Refactor to reduce duplicated code between Model2XlsxConversions impls
