+++ src/main/scala/com/norbitltd/spoiwo/natures/streaming/xlsx/Model2XlsxConversions.scala
@@ -1,66 +1,25 @@
 package com.norbitltd.spoiwo.natures.streaming.xlsx
 
 import java.io.{FileOutputStream, OutputStream}
-import java.util.{Calendar, Date}
 
 import com.norbitltd.spoiwo.model.enums._
-import com.norbitltd.spoiwo.model.{BooleanCell, CalendarCell, DateCell, FormulaCell, NoSplitOrFreeze, NumericCell, SplitPane, StringCell, _}
+import com.norbitltd.spoiwo.model.{BooleanCell, CalendarCell, DateCell, FormulaCell, NumericCell, StringCell, _}
+import com.norbitltd.spoiwo.natures.xlsx.BaseXlsx
 import com.norbitltd.spoiwo.natures.xlsx.Model2XlsxEnumConversions._
-import org.apache.poi.common.usermodel.HyperlinkType
 import org.apache.poi.ss.usermodel
 import org.apache.poi.ss.usermodel.{BorderStyle, CellType, FillPatternType, HorizontalAlignment, VerticalAlignment}
-import org.apache.poi.ss.util.{CellAddress, CellRangeAddress}
 import org.apache.poi.xssf.streaming.{SXSSFCell, SXSSFRow, SXSSFSheet, SXSSFWorkbook}
 import org.apache.poi.xssf.usermodel._
-import org.joda.time.{LocalDate, LocalDateTime}
-import org.openxmlformats.schemas.spreadsheetml.x2006.main.{CTTable, CTTableColumns, CTTableStyleInfo}
 
-object Model2XlsxConversions {
+object Model2XlsxConversions extends BaseXlsx {
 
   private type Cache[K, V] = collection.mutable.Map[SXSSFWorkbook, collection.mutable.Map[K, V]]
   private lazy val cellStyleCache = Cache[CellStyle, XSSFCellStyle]()
   private lazy val dataFormatCache = collection.mutable.Map[SXSSFWorkbook, XSSFDataFormat]()
   private lazy val fontCache = Cache[Font, XSSFFont]()
-  private val FirstSupportedDate = new LocalDate(1904, 1, 1)
-  private val LastSupportedDate = new LocalDate(9999, 12, 31)
 
   private def Cache[K, V]() = collection.mutable.Map[SXSSFWorkbook, collection.mutable.Map[K, V]]()
 
-  private def convertColor(color: Color): XSSFColor = new XSSFColor(
-    Array[Byte](color.r.toByte, color.g.toByte, color.b.toByte),
-    new DefaultIndexedColorMap()
-  )
-
-  private def mergeStyle(cell: Cell,
-                         rowStyle: Option[CellStyle],
-                         columnStyle: Option[CellStyle],
-                         sheetStyle: Option[CellStyle]): Cell = {
-    cell.styleInheritance match {
-      case CellStyleInheritance.CellOnly =>
-        cell
-      case CellStyleInheritance.CellThenRow =>
-        cell.withDefaultStyle(rowStyle)
-      case CellStyleInheritance.CellThenColumn =>
-        cell.withDefaultStyle(columnStyle)
-      case CellStyleInheritance.CellThenSheet =>
-        cell.withDefaultStyle(sheetStyle)
-      case CellStyleInheritance.CellThenColumnThenRow =>
-        cell.withDefaultStyle(columnStyle).withDefaultStyle(rowStyle)
-      case CellStyleInheritance.CellThenRowThenColumn =>
-        cell.withDefaultStyle(rowStyle).withDefaultStyle(columnStyle)
-      case CellStyleInheritance.CellThenRowThenSheet =>
-        cell.withDefaultStyle(rowStyle).withDefaultStyle(sheetStyle)
-      case CellStyleInheritance.CellThenColumnThenSheet =>
-        cell.withDefaultStyle(columnStyle).withDefaultStyle(sheetStyle)
-      case CellStyleInheritance.CellThenColumnThenRowThenSheet =>
-        cell.withDefaultStyle(columnStyle).withDefaultStyle(rowStyle).withDefaultStyle(sheetStyle)
-      case CellStyleInheritance.CellThenRowThenColumnThenSheet =>
-        cell.withDefaultStyle(rowStyle).withDefaultStyle(columnStyle).withDefaultStyle(sheetStyle)
-      case unexpected =>
-        throw new IllegalArgumentException(
-          s"Unable to convert CellStyleInheritance=$unexpected to XLSX - unsupported enum!")
-    }
-  }
 
   private[xlsx] def convertCell(modelSheet: Sheet,
                                 modelColumns: Map[Int, Column],
@@ -74,7 +33,8 @@ object Model2XlsxConversions {
     }
 
     val cellWithStyle = mergeStyle(c, modelRow.style, modelColumns.get(cellNumber).flatMap(_.style), modelSheet.style)
-    cellWithStyle.style.foreach(s => cell.setCellStyle(convertCellStyle(s, cell.getRow.getSheet.getWorkbook.asInstanceOf[SXSSFWorkbook])))
+    cellWithStyle.style.foreach(s =>
+      cell.setCellStyle(convertCellStyle(s, cell.getRow.getSheet.getWorkbook.asInstanceOf[SXSSFWorkbook])))
 
     c match {
       case BlankCell(_, _, _)               => cell.setCellValue(null: String)
@@ -89,55 +49,13 @@ object Model2XlsxConversions {
     cell
   }
 
-  private def setDateCell(c: Cell, cell: SXSSFCell, value: Date) {
-    val dateStyle = c.format.getOrElse("yyyy-MM-dd")
-    val dateTime = LocalDateTime.fromDateFields(value)
-    val date = dateTime.toLocalDate
-    if (date.isBefore(FirstSupportedDate) || date.isAfter(LastSupportedDate)) {
-      cell.setCellValue(dateTime.toString(dateStyle))
-    } else {
-      cell.setCellValue(value)
-    }
-  }
-
-  private def setCalendarCell(c: Cell, cell: SXSSFCell, value: Calendar) {
-    val dateStyle = c.format.getOrElse("yyyy-MM-dd")
-    val dateTime = LocalDateTime.fromCalendarFields(value)
-    val date = dateTime.toLocalDate
-    if (date.isBefore(FirstSupportedDate) || date.isAfter(LastSupportedDate)) {
-      cell.setCellValue(dateTime.toString(dateStyle))
-    } else {
-      cell.setCellValue(value)
-    }
-  }
-  private def setHyperLinkUrlCell(cell: SXSSFCell, value: HyperLinkUrl, row: SXSSFRow) {
-    val link = row.getSheet.getWorkbook.getCreationHelper.createHyperlink(HyperlinkType.URL)
-    link.setAddress(value.address)
-    cell.setCellValue(value.text)
-    cell.setHyperlink(link)
-
-  }
-  private[xlsx] def convertCellBorders(borders: CellBorders, style: XSSFCellStyle) {
-    borders.leftStyle.foreach(s => style.setBorderLeft(convertBorderStyle(s)))
-    borders.leftColor.foreach(c => style.setLeftBorderColor(convertColor(c)))
-    borders.bottomStyle.foreach(s => style.setBorderBottom(convertBorderStyle(s)))
-    borders.bottomColor.foreach(c => style.setBottomBorderColor(convertColor(c)))
-    borders.rightStyle.foreach(s => style.setBorderRight(convertBorderStyle(s)))
-    borders.rightColor.foreach(c => style.setRightBorderColor(convertColor(c)))
-    borders.topStyle.foreach(s => style.setBorderTop(convertBorderStyle(s)))
-    borders.topColor.foreach(c => style.setTopBorderColor(convertColor(c)))
-  }
-
-  private def convertCellDataFormat(cdf: CellDataFormat, workbook: SXSSFWorkbook, cellStyle: XSSFCellStyle) {
+  private def convertCellDataFormat(cdf: CellDataFormat, workbook: SXSSFWorkbook, cellStyle: usermodel.CellStyle) {
     cdf.formatString.foreach(formatString => {
       val format = dataFormatCache.getOrElseUpdate(workbook, workbook.getXSSFWorkbook.createDataFormat())
       val formatIndex = format.getFormat(formatString)
       cellStyle.setDataFormat(formatIndex)
     })
   }
-
-  private def convertCellRange(cr: CellRange): CellRangeAddress =
-    new org.apache.poi.ss.util.CellRangeAddress(cr.rowRange._1, cr.rowRange._2, cr.columnRange._1, cr.columnRange._2)
 
   private[xlsx] def convertCellStyle(cs: CellStyle, workbook: SXSSFWorkbook): XSSFCellStyle =
     getCachedOrUpdate(cellStyleCache, cs, workbook) {
@@ -160,103 +78,22 @@ object Model2XlsxConversions {
       cellStyle
     }
 
-  private[xlsx] def convertColumn(c: Column, sheet: SXSSFSheet) {
-    val i = c.index.getOrElse(
-      throw new IllegalArgumentException(
-        "Undefined column index! " +
-          "Something went terribly wrong as it should have been derived if not specified explicitly!"))
-
-    c.autoSized.foreach(as => if (as) sheet.autoSizeColumn(i))
-    c.break.foreach(b => if (b) sheet.setColumnBreak(i))
-    c.groupCollapsed.foreach(gc => sheet.setColumnGroupCollapsed(i, gc))
-    c.hidden.foreach(h => sheet.setColumnHidden(i, h))
-    c.width.foreach(w => sheet.setColumnWidth(i, w.toUnits))
-  }
-
-  private def convertColumnRange(cr: ColumnRange) =
-    CellRangeAddress.valueOf("%s:%s".format(cr.firstColumnName, cr.lastColumnName))
-
   private def convertFooter(f: Footer, sheet: SXSSFSheet) {
     f.left.foreach(sheet.getFooter.setLeft)
     f.center.foreach(sheet.getFooter.setCenter)
     f.right.foreach(sheet.getFooter.setRight)
-
-    /*f.firstLeft.foreach(sheet.getFirstFooter.setLeft)
-    f.firstCenter.foreach(sheet.getSheetTypeHeaderFooter.setCenter)
-    f.firstRight.foreach(sheet.getFirstFooter.setRight)*/
-
-    //Oddfooter = getFooter
-    f.oddLeft.foreach(sheet.getFooter.setLeft)
-    f.oddCenter.foreach(sheet.getFooter.setCenter)
-    f.oddRight.foreach(sheet.getFooter.setRight)
-
-/*    f.evenLeft.foreach(sheet.getEvenFooter.setLeft)
-    f.evenCenter.foreach(sheet.getEvenFooter.setCenter)
-    f.evenRight.foreach(sheet.getEvenFooter.setRight)*/
   }
 
   private[xlsx] def convertFont(f: Font, workbook: SXSSFWorkbook): XSSFFont =
     getCachedOrUpdate(fontCache, f, workbook) {
       val font = workbook.getXSSFWorkbook.createFont()
-      f.bold.foreach(font.setBold)
-      f.charSet.foreach(charSet => font.setCharSet(convertCharset(charSet).getNativeId))
-      f.color.foreach(c => font.setColor(convertColor(c)))
-      f.family.foreach(family => font.setFamily(convertFontFamily(family)))
-      f.height.foreach(height => font.setFontHeightInPoints(height.toPoints))
-      f.italic.foreach(font.setItalic)
-      f.scheme.foreach(scheme => font.setScheme(convertFontScheme(scheme)))
-      f.fontName.foreach(font.setFontName)
-      f.strikeout.foreach(font.setStrikeout)
-      f.typeOffset.foreach(offset => font.setTypeOffset(convertTypeOffset(offset)))
-      f.underline.foreach(underline => font.setUnderline(convertUnderline(underline)))
-      font
+      convertFont(f, font)
     }
 
   private def convertHeader(h: Header, sheet: SXSSFSheet) {
     h.left.foreach(sheet.getHeader.setLeft)
     h.center.foreach(sheet.getHeader.setCenter)
     h.right.foreach(sheet.getHeader.setRight)
-/*
-    h.firstLeft.foreach(sheet.getFirstHeader.setLeft)
-    h.firstCenter.foreach(sheet.getFirstHeader.setCenter)
-    h.firstRight.foreach(sheet.getFirstHeader.setRight)*/
-
-    h.oddLeft.foreach(sheet.getHeader.setLeft)
-    h.oddCenter.foreach(sheet.getHeader.setCenter)
-    h.oddRight.foreach(sheet.getHeader.setRight)
-
-/*    h.evenLeft.foreach(sheet.getEvenHeader.setLeft)
-    h.evenCenter.foreach(sheet.getEvenHeader.setCenter)
-    h.evenRight.foreach(sheet.getEvenHeader.setRight)*/
-  }
-
-  private def convertMargins(margins: Margins, sheet: SXSSFSheet) {
-    margins.top.foreach(topMargin => sheet.setMargin(usermodel.Sheet.TopMargin, topMargin))
-    margins.bottom.foreach(bottomMargin => sheet.setMargin(usermodel.Sheet.BottomMargin, bottomMargin))
-    margins.right.foreach(rightMargin => sheet.setMargin(usermodel.Sheet.RightMargin, rightMargin))
-    margins.left.foreach(leftMargin => sheet.setMargin(usermodel.Sheet.LeftMargin, leftMargin))
-    margins.header.foreach(headerMargin => sheet.setMargin(usermodel.Sheet.HeaderMargin, headerMargin))
-    margins.footer.foreach(footerMargin => sheet.setMargin(usermodel.Sheet.FooterMargin, footerMargin))
-  }
-
-  private def convertPane(pane: Pane): Int = pane match {
-    case Pane.LowerLeftPane  => org.apache.poi.ss.usermodel.Sheet.PANE_LOWER_LEFT
-    case Pane.LowerRightPane => org.apache.poi.ss.usermodel.Sheet.PANE_LOWER_RIGHT
-    case Pane.UpperLeftPane  => org.apache.poi.ss.usermodel.Sheet.PANE_UPPER_LEFT
-    case Pane.UpperRightPane => org.apache.poi.ss.usermodel.Sheet.PANE_UPPER_RIGHT
-    case unexpected =>
-      throw new IllegalArgumentException(s"Unable to convert Pane=$unexpected to XLSX - unsupported enum!")
-  }
-
-  private def convertPaneAction(paneAction: PaneAction, sheet: SXSSFSheet) {
-    paneAction match {
-      case NoSplitOrFreeze() =>
-        sheet.createFreezePane(0, 0)
-      case FreezePane(columnSplit, rowSplit, leftMostColumn, topRow) =>
-        sheet.createFreezePane(columnSplit, rowSplit, leftMostColumn, topRow)
-      case SplitPane(xSplitPosition, ySplitPosition, leftMostColumn, topRow, activePane) =>
-        sheet.createSplitPane(xSplitPosition, ySplitPosition, leftMostColumn, topRow, convertPane(activePane))
-    }
   }
 
   private[xlsx] def convertRow(modelColumns: Map[Int, Column],
@@ -271,18 +108,6 @@ object Model2XlsxConversions {
     modelRow.hidden.foreach(row.setZeroHeight)
     modelRow.cells.foreach(cell => convertCell(modelSheet, modelColumns, modelRow, cell, row))
     row
-  }
-
-  private def validateCells(r: Row) {
-    val indexedCells = r.cells.filter(_.index.isDefined)
-    val contextCells = r.cells.filter(_.index.isEmpty)
-
-    if (indexedCells.nonEmpty && contextCells.nonEmpty)
-      throw new IllegalArgumentException("It is not allowed to mix cells with and without index within a single row!")
-
-    val distinctIndexes = indexedCells.map(_.index).toSet.flatten
-    if (indexedCells.size != distinctIndexes.size)
-      throw new IllegalArgumentException("It is not allowed to have cells with duplicate index within a single row!")
   }
 
   private[xlsx] def convertSheet(s: Sheet, workbook: SXSSFWorkbook): SXSSFSheet = {
@@ -309,156 +134,25 @@ object Model2XlsxConversions {
     s.repeatingColumns.foreach(rc => sheet.setRepeatingColumns(convertColumnRange(rc)))
     s.password.foreach(ps => sheet.protectSheet(ps))
 
-    if(!s.tables.isEmpty) {
-      throw  new IllegalStateException("createTable is not supported by SXSSF right now")
-    }
-
     sheet
   }
-  private def validateRows(s: Sheet) {
-    val indexedRows = s.rows.filter(_.index.isDefined)
-    val contextRows = s.rows.filter(_.index.isEmpty)
 
-    if (indexedRows.nonEmpty && contextRows.nonEmpty) {
-      throw new IllegalArgumentException("It is not allowed to mix rows with and without index within a single sheet!")
-    }
-
-    val distinctIndexes = indexedRows.flatMap(_.index).toSet
-    if (indexedRows.size != distinctIndexes.size)
-      throw new IllegalArgumentException("It is not allowed to have rows with duplicate index within a single sheet!")
+  override def setTabColor(sheet: usermodel.Sheet, color: XSSFColor) {
+    sheet.asInstanceOf[SXSSFSheet].setTabColor(color)
   }
 
-  private def updateColumnsWithIndexes(s: Sheet): List[Column] = {
-    val currentColumnIndexes = s.columns.flatMap(_.index).toSet
-    if (currentColumnIndexes.isEmpty) {
-      s.columns.zipWithIndex.map {
-        case (column, index) => column.withIndex(index)
-      }
-    } else if (currentColumnIndexes.size == s.columns.size) {
-      s.columns
-    } else {
-      throw new IllegalArgumentException(
-        "When explicitly specifying column index you are required to provide it " +
-          "uniquely for all columns in this sheet definition!")
-    }
-  }
-
-  private[xlsx] def convertSheetProperties(sp: SheetProperties, sheet: SXSSFSheet) {
-    sp.autoFilter.foreach(autoFilterRange => sheet.setAutoFilter(convertCellRange(autoFilterRange)))
-    sp.activeCell.foreach(stringReference => sheet.setActiveCell(new CellAddress(stringReference)))
-    sp.autoBreaks.foreach(sheet.setAutobreaks)
-    sp.defaultColumnWidth.foreach(sheet.setDefaultColumnWidth)
-    sp.defaultRowHeight.foreach(height => sheet.setDefaultRowHeightInPoints(height.toPoints))
-    sp.displayFormulas.foreach(sheet.setDisplayFormulas)
-    sp.displayGridLines.foreach(sheet.setDisplayGridlines)
-    sp.displayGuts.foreach(sheet.setDisplayGuts)
-    sp.displayRowColHeadings.foreach(sheet.setDisplayRowColHeadings)
-    sp.displayZeros.foreach(sheet.setDisplayZeros)
-    sp.fitToPage.foreach(sheet.setFitToPage)
-    sp.forceFormulaRecalculation.foreach(sheet.setForceFormulaRecalculation)
-    sp.horizontallyCenter.foreach(sheet.setHorizontallyCenter)
-    sp.printArea.foreach {
-      case CellRange((startRow, endRow), (startColumn, endColumn)) =>
-        val workbook = sheet.getWorkbook
-        val sheetIndex = workbook.getNumberOfSheets - 1
-        workbook.setPrintArea(sheetIndex, startColumn, endColumn, startRow, endRow)
-    }
-    sp.printGridLines.foreach(sheet.setPrintGridlines)
-    sp.rightToLeft.foreach(sheet.setRightToLeft)
-    sp.rowSumsBelow.foreach(sheet.setRowSumsBelow)
-    sp.rowSumsRight.foreach(sheet.setRowSumsRight)
-    sp.selected.foreach(sheet.setSelected)
-    sp.tabColor.foreach(c => sheet.setTabColor(convertColor(c)))
-    sp.virtuallyCenter.foreach(sheet.setVerticallyCenter)
-    sp.zoom.foreach(sheet.setZoom)
-  }
-
-  private def convertPrintSetup(printSetup: PrintSetup, sheet: SXSSFSheet) {
+  override def additionalPrintSetup(printSetup: PrintSetup, sheetPs: usermodel.PrintSetup) {
     if (printSetup != PrintSetup.Default) {
-      val ps = sheet.getPrintSetup.asInstanceOf[XSSFPrintSetup]
-      printSetup.copies.foreach(ps.setCopies)
-      printSetup.draft.foreach(ps.setDraft)
-      printSetup.fitHeight.foreach(ps.setFitHeight)
-      printSetup.fitWidth.foreach(ps.setFitWidth)
-      printSetup.footerMargin.foreach(ps.setFooterMargin)
-      printSetup.headerMargin.foreach(ps.setHeaderMargin)
-      printSetup.hResolution.foreach(ps.setHResolution)
-      printSetup.landscape.foreach(ps.setLandscape)
-      printSetup.leftToRight.foreach(ps.setLeftToRight)
-      printSetup.noColor.foreach(ps.setNoColor)
-      printSetup.noOrientation.foreach(ps.setNoOrientation)
+      val ps = sheetPs.asInstanceOf[XSSFPrintSetup]
       printSetup.pageOrder.foreach(po => ps.setPageOrder(convertPageOrder(po)))
-      printSetup.pageStart.foreach(ps.setPageStart)
       printSetup.paperSize.foreach(px => ps.setPaperSize(convertPaperSize(px)))
-      printSetup.scale.foreach(ps.setScale)
-      printSetup.usePage.foreach(ps.setUsePage)
-      printSetup.validSettings.foreach(ps.setValidSettings)
-      printSetup.vResolution.foreach(ps.setVResolution)
     }
   }
-
-  private def convertRowRange(rr: RowRange) =
-    CellRangeAddress.valueOf("%d:%d".format(rr.firstRowIndex, rr.lastRowIndex))
 
   private[xlsx] def validateTables(modelSheet: Sheet): Unit = {
-    val ids = modelSheet.tables.flatMap(_.id)
-    if (ids.size != ids.toSet.size)
-      throw new IllegalArgumentException("Specified table ids need to be unique.")
-  }
-
-  private[xlsx] def validateTableColumns(modelTable: Table): Unit = {
-
-    def insufficientColumnsDefined = {
-      val (sCol, eCol) = modelTable.cellRange.columnRange
-      val neededColumns = (eCol - sCol) + 1
-      val definedColumns = modelTable.columns.size
-      neededColumns != definedColumns
+    if (!modelSheet.tables.isEmpty) {
+      throw new IllegalStateException("createTable is not supported by SXSSF right now")
     }
-
-    if (modelTable.columns.nonEmpty && insufficientColumnsDefined) {
-      throw new IllegalArgumentException(
-        "When explicitly specifying table columns you are required to provide as many " +
-          "columns as in the cell range."
-      )
-    }
-  }
-
-  private[xlsx] def convertTableColumns(modelTable: Table, ctTable: CTTable): CTTableColumns = {
-
-    def generateColumns = {
-      val (sCol, eCol) = modelTable.cellRange.columnRange
-      val neededColumns = (eCol - sCol) + 1
-      (0 until neededColumns) map { index ⇒
-        val columnId = index + 1
-        TableColumn(
-          name = s"TableColumn$columnId",
-          id = columnId.toLong
-        )
-      }
-    }
-
-    val modelColumns = if (modelTable.columns.nonEmpty) modelTable.columns else generateColumns
-    val columns = ctTable.addNewTableColumns()
-    columns.setCount(modelColumns.size)
-    modelColumns.foreach { mc ⇒
-      val column = columns.addNewTableColumn()
-      column.setName(mc.name)
-      column.setId(mc.id)
-    }
-    columns
-  }
-
-  private[xlsx] def convertTableStyle(modelTableStyle: TableStyle, ctTable: CTTable): CTTableStyleInfo = {
-    val style = ctTable.addNewTableStyleInfo()
-    style.setName(modelTableStyle.name.value)
-    modelTableStyle.showColumnStripes.foreach(style.setShowColumnStripes)
-    modelTableStyle.showRowStripes.foreach(style.setShowRowStripes)
-    style
-  }
-
-  private[xlsx] def setTableReference(modelTable: Table, ctTable: CTTable): Unit = {
-    val cellRangeAddress = convertCellRange(modelTable.cellRange)
-    ctTable.setRef(cellRangeAddress.formatAsString())
   }
 
   private[xlsx] def convertWorkbook(wb: Workbook): SXSSFWorkbook = {
